login() function and user_login_tracker()
Inside the the MockATMProject.py, I added more features to my login function that has the ability to track when a user is logged in an when the user logs in a new file is created in the auth session folder which holds the user information.  
I also created a similar function called user_login tracker inside the database.py file as well. 
Also incorporated a a global balance so whenever a user conducts transactions, the balance does not reset.